# app-connector-client/server/api.js
var CONFIG = require("./config")
const path = require("path")
var request = require("request")
var LOGGER = require("./logger")
const fs = require("fs")

const keyFile = path.resolve(CONFIG.keyDir, 'app.key')
const certFile = path.resolve(CONFIG.keyDir, 'kyma.crt')

exports.getAll = function (req, res) {
    LOGGER.logger.debug("Getting all APIs")

    module.exports.getAPIs(req.query.localKyma == true, function (data, err) {
        if (err) {
            LOGGER.logger.error("Error while getting all APIs: %s", err)
            res.status(500).send(err)
        } else if (!data) {
            res.status(200).send([])
        } else {
            try {
                LOGGER.logger.debug("Received API data: %s", data)
                res.status(200).send(JSON.parse(data))
            }
            catch (err2) {
                LOGGER.logger.error("Error while parsing response payload: %s", err2)
                res.status(200).send(data)
            }
        }
    })
};

exports.create = function (req, res) {
    LOGGER.logger.debug("Creating API")

    module.exports.createAPI(req.query.localKyma == true, req.body, function (data, err) {
        if (err) {
            LOGGER.logger.error("Error while creating API: %s", err)
            res.status(500).send(err)
        } else if (!data) {
            res.status(200).send({})
        } else {
            try {
                LOGGER.logger.debug("Received API data: %s", data)
                res.status(200).send(JSON.parse(data))
            }
            catch (err2) {
                LOGGER.logger.error("Error while parsing response payload: %s", err2)
                res.status(200).send(data)
            }
        }
    })
};

exports.get = function (req, res) {
    LOGGER.logger.debug("Get API %s", req.params.api)

    module.exports.getAPI(req.query.localKyma == true, req.params.api, function (data, err) {
        if (err) {
            LOGGER.logger.error("Error while getting API: %s", err)
            res.status(500).send(err)
        } else if (!data) {
            res.status(200).send({})
        } else {
            try {
                LOGGER.logger.debug("Received API data: %s", data)
                res.status(200).send(JSON.parse(data))
            }
            catch (err2) {
                LOGGER.logger.error("Error while parsing response payload: %s", err2)
                res.status(200).send(data)
            }
        }
    })
};

exports.update = function (req, res) {
    LOGGER.logger.debug("Update API %s", req.params.api)

    module.exports.updateAPI(req.query.localKyma == true, req.params.api, req.body, function (data, err) {
        if (err) {
            LOGGER.logger.error("Error while updating API: %s", err)
            res.status(500).send(err)
        } else if (!data) {
            res.status(200).send({})
        } else {
            try {
                LOGGER.logger.debug("Received API data: %s", data)
                res.status(200).send(JSON.parse(data))
            }
            catch (err2) {
                LOGGER.logger.error("Error while parsing response payload: %s", err2)
                res.status(200).send(data)
            }
        }
    })
};

exports.delete = function (req, res) {
    LOGGER.logger.debug("Delete API %s", req.params.api)

    module.exports.deleteAPI(req.query.localKyma == true, req.params.api, req.body, function (data, err) {
        if (err) {
            LOGGER.logger.error("Error while deleting API: %s", err)
            res.status(500).send(err)
        } else {
            LOGGER.logger.debug("Received API data: %s", data)
            res.status(204).send()
        }
    })
};

exports.deleteAll = function (req, res) {
    err = assureConnected()
    if (err) {
        cb(null, err)
    } else {
        LOGGER.logger.debug("Creating API with payload: %s", payload)

        module.exports.deleteAPIs(req.query.localKyma == true, req.params.api, function (data, err) {
            if (err) {
                LOGGER.logger.error("Error while deleting APIs: %s", err)
                res.status(500).send(err)
            } else {
                res.status(204).send()
            }
        })
    }
};

exports.createAPI = function createAPI(localKyma, payload, cb) {
    err = assureConnected()
    if (err) {
        cb(null, err)
    } else {
        LOGGER.logger.debug("Creating API with payload: %s", payload)

        request.post({
            url: CONFIG.URLs.metadataUrl,
            headers: {
                "Content-Type": "application/json"
            },
            json: payload,
            agentOptions: {
                cert: fs.readFileSync(certFile),
                key: fs.readFileSync(keyFile)
            },
            rejectUnauthorized: !localKyma
        }, function (error, httpResponse, body) {
            if (error) {
                cb(null, error)
            }
            cb(body, null)
        });
    }
}

exports.deleteAPI = function deleteAPI(localKyma, id, cb) {
    err = assureConnected()
    if (err) {
        cb(null, err)
    } else {
        LOGGER.logger.debug("Deleting API: %s", id)
        request.delete(
            {
                url: `${CONFIG.URLs.metadataUrl}/${id}`,
                agentOptions: {
                    cert: fs.readFileSync(certFile),
                    key: fs.readFileSync(keyFile)
                },
                rejectUnauthorized: !localKyma
            }, function (error, httpResponse, body) {
                if (error) {
                    cb(null, error)
                }
                cb(body, null)
            }
        )
    }
}

exports.deleteAPIs = function deleteAPIs(localKyma, cb) {
    err = assureConnected()
    if (err) {
        cb(null, err)
    } else {
        LOGGER.logger.debug("Deleting all APIs")
        request.delete(
            {
                url: `${CONFIG.URLs.metadataUrl}`,
                agentOptions: {
                    cert: fs.readFileSync(certFile),
                    key: fs.readFileSync(keyFile)
                },
                rejectUnauthorized: !localKyma
            }, function (error, httpResponse, body) {
                if (error) {
                    cb(null, error)
                }
                cb(body, null)
            }
        )
    }
}

exports.getAPIs = function getAPIs(localKyma, cb) {
    err = assureConnected()
    if (err) {
        cb(null, err)
    } else {
        LOGGER.logger.debug("Retrieving all APIs")
        request({
            url: CONFIG.URLs.metadataUrl,
            method: "GET",
            agentOptions: {
                cert: fs.readFileSync(certFile),
                key: fs.readFileSync(keyFile)
            },
            rejectUnauthorized: !localKyma
        }, function (error, httpResponse, body) {
            if (error) {
                cb(null, error)
            }
            cb(body, null)
        })
    }
}

exports.getAPI = function getAPI(localKyma, id, cb) {
    err = assureConnected()
    if (err) {
        cb(null, err)
    } else {
        LOGGER.logger.debug("Retrieving API: %s", id)
        request.get({
            url: `${CONFIG.URLs.metadataUrl}/${id}`,
            agentOptions: {
                cert: fs.readFileSync(certFile),
                key: fs.readFileSync(keyFile)
            },
            rejectUnauthorized: !localKyma
        }, function (error, httpResponse, body) {
            if (error) {
                cb(null, error)
            }
            cb(body, null)
        })
    }
}

exports.updateAPI = function updateAPI(localKyma, id, payload, cb) {
    err = assureConnected()
    if (err) {
        cb(null, err)
    } else {
        LOGGER.logger.debug("Updating API: %s", id)
        request.put({
            url: `${CONFIG.URLs.metadataUrl}/${id}`,
            headers: {
                "Content-Type": "application/json"
            },
            json: payload,
            agentOptions: {
                cert: fs.readFileSync(certFile),
                key: fs.readFileSync(keyFile)
            },
            rejectUnauthorized: !localKyma
        }, function (error, httpResponse, body) {
            if (error) {
                cb(null, error)
            }
            cb(body, null)
        });
    }
}

function assureConnected() {
    if (CONFIG.URLs.metadataUrl == "") {
        return "Not connected to a kyma cluster, please re-connect"
    }
    return null
}