# app-connector-client/server/resources/swagger.yaml
openapi: '3.0.0'
info:
  version: '1.0.0'
  title: 'Kyma Metadata API'
tags:
- name: 'Metadata Service'
  description: 'API for registering APIs in Kyma.'
paths:
  /metadata:
    get:
      tags:
        - 'metadata'
      summary: 'Returns this swagger definition'
      responses:
        '200':
          description: "Successful"
          content:
            application/yaml:
              type: yaml         
  /connection:
    post:
      tags:
        - 'connection'
      summary: "Sends your token and connects you to Kyma"
      parameters:
      - in: 'query'
        name: 'localKyma'
        description: 'deployment to local kyma'
        required: true
        schema:
          type: 'boolean'
      responses:
       '200':
        description: 'Successful connection to Kyma'
        content:
          application/json:
            schema:
              type: 'object'
              properties:
                metadataUrl:
                  type: 'string'
                eventsUrl:
                  type: 'string'
                certificatesUrl:
                  type: 'string'
       '401':
          description: 'Connection is not authorized'
          content:
            text/plain:
              schema:
                type: 'string'
                example: There is an error while registering. Please make sure that your token is unique
    get:
      tags:
        - "connection"
      summary: Returns active connection info
      responses:
        '200':
          description: "Currently connected remote env & cluster"
          content:
            application/json:
              schema:
                type: object
                properties:
                  cluster_domain:
                    type: string
                  re_name:
                    type: string
                  gateway_url:
                    type: string
  /connection/key:
    get:
      tags:
        - 'certificates'
      summary: 'Get your private key'
      responses:
        '200':
          description: your private key
          content:
            application/file:
             schema:
               type: 'string'
               format: binary
  /connection/cert:
    get:
      tags:
        - 'certificates'
      summary: Get certificate returned by kyma
      responses:
        '200':
          description: your kyma certificate
          content:
            application/file:
             schema:
               type: 'string'
               format: binary
  /events:
    post:
      tags:
      - 'events'
      summary: 'Sends a new event'
      operationId: 'sendEvent'
      requestBody:
        description: 'Event payload'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventDetails'
      parameters:
      - in: 'query'
        name: 'localKyma'
        description: 'deployment to local kyma'
        required: true
        schema:
          type: 'boolean'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceId'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataErrorResponse'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataErrorResponse'
  /apis:
    post:
      tags:
      - 'apis'
      summary: 'Registers a new API'
      operationId: 'createAPI'
      requestBody:
        description: 'API to be registered'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceDetails'
      parameters:
      - in: 'query'
        name: 'localKyma'
        description: 'deployment to local kyma'
        required: true
        schema:
          type: 'boolean'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceId'
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataErrorResponse'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataErrorResponse'
    get:
      tags:
      - 'apis'
      summary: 'Gets all registered apis'
      operationId: 'getAPIs'
      parameters:
      - in: 'query'
      name: 'localKyma'
      description: 'deployment to local kyma'
      required: true
      schema:
        type: 'boolean'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                type: 'array'
                items:
                  $ref: '#/components/schemas/Service'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataErrorResponse'
  /apis/{Id}:
    get:
      tags:
      - 'apis'
      summary: 'Gets a API by ID'
      operationId: 'getAPI'
      parameters:
      - in: 'path'
        name: 'id'
        description: 'ID of a API'
        required: true
        schema:
          type: 'string'
          format: 'uuid'
      - in: 'query'
        name: 'localKyma'
        description: 'deployment to local kyma'
        required: true
        schema:
          type: 'boolean'
      responses:
        '200':
          description: 'Successful operation'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetails'
        '404':
          description: 'API not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataErrorResponse'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataErrorResponse'
    put:
      tags:
      - 'apis'
      summary: 'Updates a API by ID'
      operationId: 'updateAPI'
      requestBody:
        description: 'API to be stored'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceDetails'
      parameters:
      - in: 'path'
        name: 'Id'
        description: 'ID of a service'
        required: true
        schema:
          type: 'string'
          format: 'uuid'
      - in: 'query'
        name: 'localKyma'
        description: 'deployment to local kyma'
        required: true
        schema:
          type: 'boolean'
      responses:
        '200':
          description: 'Successful operation'
        '404':
          description: 'API not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataErrorResponse'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataErrorResponse'
    delete:
      tags:
      - 'apis'
      summary: 'Deletes a API by ID'
      operationId: 'deleteAPI'
      parameters:
      - in: 'path'
        name: 'Id'
        description: 'ID of a API'
        required: true
        schema:
          type: 'string'
          format: 'uuid'
      - in: 'query'
        name: 'localKyma'
        description: 'deployment to local kyma'
        required: true
        schema:
          type: 'boolean'
      responses:
        '204':
          description: 'Successful operation'
        '404':
          description: 'API not found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataErrorResponse'
        '500':
          description: 'Internal server error'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataErrorResponse'
components:
  schemas:
    ServiceId:
      type: 'object'
      properties:
        id:
          type: 'string'
          format: 'uuid'
    ServiceDetails:
      type: 'object'
      properties:
        provider:
          type: 'string'
        name:
          type: 'string'
        description:
          type: 'string'
        api:
          $ref: '#/components/schemas/Api'
        events:
          $ref: '#/components/schemas/Events'
        documentation:
          $ref: '#/components/schemas/Documentation'
      required:
        - provider
        - name
        - description
    Service:
      type: 'object'
      properties:
        id:
          type: 'string'
          format: 'uuid'
        provider:
          type: 'string'
        name:
          type: 'string'
        description:
          type: 'string'
    Api:
      type: 'object'
      properties:
        targetUrl:
          type: 'string'
          format: 'uri'
        credentials:
          $ref: '#/components/schemas/ApiCredentials'
        spec:
          type: 'object'
          description: 'OpenApi v2 swagger file: https://github.com/OAI/OpenAPI-Specification/blob/master/schemas/v2.0/schema.json'
      required:
        - targetUrl
    Events:
      type: 'object'
      properties:
        spec:
          description: 'AsynApi file v1: https://github.com/asyncapi/asyncapi/blob/develop/schema/asyncapi.json'
          type: 'object'
    Documentation:
      type: 'object'
      properties:
        displayName: 
          type: 'string'
        description: 
          type: 'string'
        type:
          type: 'string'
        tags:
          type: 'array'
          items: 
            type: 'string'
        docs:
          type: 'array'
          items:
            $ref: '#/components/schemas/Document'
      required:
        - displayName
        - description
        - type
    Document:
      type: 'object'
      properties:
        title:
          type: 'string'
        type:
          type: 'string'
        source:
          type: 'string'
      required:
        - title
        - type
        - source
    ApiCredentials:
      type: 'object'
      properties:
        oauth:
          $ref: '#/components/schemas/OAuth'
      required:
        - oauth
    OAuth:
      type: 'object'
      properties:
        url:
          type: 'string'
          format: 'uri'
        clientId:
          type: 'string'
        clientSecret:
          type: 'string'
      required:
        - url
        - clientId
        - clientSecret
    MetadataErrorResponse:
      type: 'object'
      properties:
        code:
          type: 'integer'
        error:
          type: 'string'